move MCMC-specific test from test_methods to test_mcmc
